move ITreeItemData to its own file
